+++ CHANGELOG.md
@@ -8,8 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - removed description and discussion from Logo structure
 - removed isDistributed from Backer structure
-- added logoRewards variable to avoid loops in dLogos.sol
-- added logoRejectedFunds variable to avoid loops in dLogos.sol
+### Added
+- added logoRewards variable to avoid loops in Dlogos
+- added logoRejectedFunds variable to avoid loops in Dlogos
+- added pause() and unpause() in Dlogos
 
 ## [2.0.1] - 2024-06-03
 ### Added